js: Add `type(...)` expressions and allow scoped blocks in statements again
Since `type` is a keyword, I decided to add a special case by making a 
new type of expression for `type(...)` calls, which return an 
implementation-defined string representation of the N type.

Blocks are used as a collection of Statements, but they used to create 
their own scope which messed up file exporting. However, they're also 
used as a statement to contain a scope:

{
  ...
}

Now, WrappedBlocks handle that instead of Blocks